==> 01a_ETL_games.ipynb <==
<h2 align=center>  Proyecto Integrador I STEAM</h2>

En esta notebook llevaremos a cabo el proceso de ETL (Extraction, Transformation & Load) para el archivo output_steam_games.json que se encuentra en la carpeta de data. Procederemos a leerlo, analizar nulos, registros vacios y de ser necesario eliminarlos o inputar valores, para terminar cargando los datos nuevamente a un archivo csv con la información limpia.

### Extraccion de datos steam_games

Importamos las librerias a utilizar

import pandas as pd
import numpy as np
import json
import warnings
import Tools as t
warnings.filterwarnings("ignore")

Extraemos los datos del dataset y los almacenamos en un dataframe

ruta_games = 'data/output_steam_games.json'

filas_games = []                        # Va linea por linea del json
with open(ruta_games) as f:
    for line in f.readlines():
        data = json.loads(line)         # carga cada linea
        filas_games.append(data)        # y cada registro lo agrega a la lista filas_games, asi hasta terminar todos los registros
df_juegos = pd.DataFrame(filas_games)   # pasamos los registros a un dataframe
df_juegos

### Transformacion de datos

Vemos que tipo de dato es cada variable, la cantidad de registros y columnas

df_juegos.info()

Veamos que hay en las columnas tags y specs

df_juegos['tags'].value_counts()

df_juegos['specs'].value_counts()

Calculamos la cantidad de registros cuyos datos son en su totalidad nulos, y procedemos a eliminarlos

cantidad_filas_nulas = df_juegos.isnull().all(axis=1).sum()                             # Cantidad de filas con valores nulas
print("Cantidad de filas con todos los valores nulos:", cantidad_filas_nulas)
if cantidad_filas_nulas > 0:                                                            # Si es >0, los elimina
    df_juegos2 = df_juegos.dropna(how='all').reset_index(drop=True)
    print('Se eliminaron los registros totalmente vacios')
else:
    print('No se eliminó registros, ya que no se encontraron registros totalmente vacios')

Luego buscamos cuantos nulos quedan en nuestro dataframe, para esto utilizamos una funcion llamada PorcentajeNulos que se encuentra en el archivo Tools.py

nulos = t.PorcentajeNulos(df_juegos2)
nulos 

Como hay una cantidad considerable de nulos, por el momento no eliminaremos nada hasta más adelante

Por fines meramente de optimización de tiempo, procederemos a eliminar los valores nulos de las columna developer, en una fase posterior a este análisis el plan es rellenar esos nulos con webscrapping 

df_juegos2 = df_juegos2.dropna(subset=['developer'], axis=0)

Para simplificar el dataframe, validamos si la columna app name y title contienen la misma informacion

df = df_juegos2

df['Comparacion'] = 'Distintos'                                                                 # Comprobar si son distintas
df.loc[(df['app_name'].isnull()) & (df['title'].isnull()), 'Comparacion'] = 'Ambos nulos'       # Comprobar si son nulos
df.loc[df['app_name'] == df['title'], 'Comparacion'] = 'Iguales'                                # Comprobar si ambas columnas son iguales

# Calcular la suma de valores distintos, iguales y nulos
suma_valores = df['Comparacion'].value_counts()                                                                 

suma_valores

Revisamos varios registros al azar para validar que sean la misma información

registro_distinto = df[df['Comparacion'] == 'Distintos'].iloc[478]                              # Validar si los datos distintos realmente son distintos para eliminarlos posteriormente
registro_distinto

Por lo que procedemos a eliminar la columna "title"; también eliminamos la columna "tags" porque hay informacion repetida que además no es relevante para los endpoints posteriores, al igual que las columnas "url", "reviews_url" y "specs".

eliminar = ['title','url','reviews_url','tags','specs','Comparacion', 'price','publisher','early_access'] 
df_juegos2 = df_juegos2.drop(columns=eliminar)
df_juegos2.head()

Desanidamos la columna genres

df_juegos3 = df_juegos2.explode('genres')               # explode() generará una nueva fila por cada elemento de la lista
df_juegos3.head()

Para simplificar el dataframe, agregaremos la columna released_year la cual contiene únicamente el año de lanzamiento ya que en el endpoint no es necesario día y mes. A su vez, completamos los datos nulos con la mediana de los años de los demás registros

df_juegos3['released_year'] = df_juegos3['release_date'].str.extract(r'(\d{4})')            # extraemos el año
media = df_juegos3['released_year'].dropna().astype(float).median()                         # como son más de 10,000 datos que no cuentan con el dato del año
df_juegos3['released_year'] = df_juegos3['released_year'].fillna(media)                     # rellenamos esos nulos con la media de los años con de los que si tienen el dato del year_released
del df_juegos3['release_date']

df_juegos3.head()

Ya que contamos con el dataframe limpio, procedemos a modificar el tipo de dato de cada columna

df_juegos3['genres'] = df_juegos3['genres'].astype(str)
df_juegos3['app_name'] = df_juegos3['app_name'].astype(str)
df_juegos3['developer'] = df_juegos3['developer'].astype(str)
df_juegos3['released_year'] = df_juegos3['released_year'].astype(int)

Cambiamos el nombre de la columna id por id_item, para poder juntarla con otras columnas más adelante

df_juegos3.rename(columns={'id':'item_id'}, inplace = True)     

### Carga de datos. 

# Guardamos el conjunto ya limpio a un csv
df_juegos3.to_csv('data_ok/games_cleaned.csv', index = False, encoding='utf-8')

Este es el fin de este ETL, porfavor da click [aqui](01b_ETL_items.ipynb) para continuar.
