==> 01b_ETL_items.ipynb <==
En esta notebook llevaremos a cabo el proceso de ETL para el archivo australian_users_items.json que se encuentra en la carpeta de data. Procederemos a leerlo, analizar nulos, registros vacios y de ser necesario eliminarlos o inputar valores, para terminar cargando los datos nuevamente a un archivo csv con la información limpia.

### Extracción de datos Users Items

import pandas as pd
import numpy as np
import json
import Tools as t
import warnings
warnings.filterwarnings("ignore")

Extraemos los datos del dataset, desanidamos columnas y los almacenamos en un dataframe

ruta_items = 'data/australian_users_items.json'
import ast

filas_items = []

with open(ruta_items, encoding='utf-8') as f:
    for line in f.readlines():
        try:
            data = json.loads(line)                                                                         # Intenta cargar el JSON normalmente
        except json.JSONDecodeError:                                                                        # Si tiene error:
            data = ast.literal_eval(line)                                                                   # Usa ast.literal_eval() para cargar el JSON inválido y cambiar '' por ""
        filas_items.append(data)                                                                            # Se agrega a una lista_items

df_items = pd.DataFrame(filas_items)                                                                        # Se pasa a un dataframe        
data_user_item1 = df_items.explode(['items'])                                                               # Desanidamos la columna items
data_user_item2 = pd.json_normalize(data_user_item1['items']).set_index(data_user_item1['items'].index)     # "Aplanamos" el dataframe
df_items2= pd.concat([data_user_item2, data_user_item1], axis=1)                                            # Unimos los df
df_items2

Eliminamos la columna items, ya que ya fue desanidada y sus elementos anexados al dataframe

del df_items2['items']

#### Transformación de datos

Vemos que tipo de dato es cada variable, la cantidad de registros y columnas

df_items2.info()

Calculamos la cantidad de valores nulos en el dataframe df_items2

nulos = t.PorcentajeNulos(df_items2)
nulos

Eliminamos los registros nulos, ya que no son una cantidad significativa (menos del 2% del total)

df_items2 = df_items2.dropna()

Revisamos los duplicados

duplicados = df_items2[df_items2.duplicated(keep=False)]
duplicados.head()

Al verificar si son resultados duplicados, observamos que las horas de playtime_forever y item_name son distintas a pesar que todo lo demás es igual, por lo que no se eliminarán porque no son realmente duplicados.

user_id = '76561198156664158'

fila_usuario = df_items2.loc[df_items2['user_id'] == user_id]
fila_usuario.head()

Eliminaremos las columnas que no se utilizarán en los endpoints: playtime_2weeks, user_url

eliminar = ['playtime_2weeks', 'user_url']
df_items2 = df_items2.drop(columns=eliminar)

Ya que contamos con el dataframe limpio, proseguimos con la modificación del tipo de dato de cada columna

df_items2['user_id'] = df_items2['user_id'].astype(str)
df_items2['item_name'] = df_items2['item_name'].astype(str)
df_items2['user_id'] = df_items2['user_id'].astype(str)

### Carga de datos

# Guardamos el conjunto ya limpio a un csv
df_items2.to_csv('users_items_cleaned.csv', index = False, encoding='utf-8')

Este es el fin de este ETL, porfavor da click [aqui](01c_ETL_reviews.ipynb) para continuar.
