==> 01c_ETL_reviews.ipynb <==
En esta notebook llevaremos a cabo el proceso de ETL para el archivo australian_user_reviews.json que se encuentra en la carpeta de data. Procederemos a leerlo, analizar nulos, registros vacios y de ser necesario eliminarlos o inputar valores, para terminar cargando los datos nuevamente a un archivo csv con la información limpia.

### Extracción de datos users_reviews

Importamos las bibliotecas a utilizar

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import json
import warnings
from datetime import datetime
import Tools as t
warnings.filterwarnings("ignore")

Extraemos los datos de los datasets

ruta_review = 'data/australian_user_reviews.json'
import ast

filas_review = []

with open(ruta_review, encoding='utf-8') as f:
    for line in f.readlines():
        try:
            data = json.loads(line)         # Intenta cargar el JSON normalmente
        except json.JSONDecodeError:        # Si tiene error: 
            data = ast.literal_eval(line)   # Usa ast.literal_eval() para cargar el JSON inválido y cambiar '' por ""
        filas_review.append(data)           # Agrega la linea corregida a la lista

df_review = pd.DataFrame(filas_review)      # Una vez cargados todas las lineas a la lista, hace un DataFrame con dicha lista
df_review                                   # validamos el DataFrame

De la columna reviews, sacamos la informacion de los diccionarios

df_clean_reviews = df_review.explode('reviews')                                                                                 # Desanidamos la columna reviews

'''
Elimina la columna 'reviews' del DataFrame df_clean_reviews y expande la columna 'reviews' en varios campos (si cada elemento de la columna 'reviews' contiene varios valores) para luego 
unir esta información en un nuevo DataFrame df_reviews2
'''

df_reviews2 = pd.concat([df_clean_reviews.drop(['reviews'], axis=1), df_clean_reviews['reviews'].apply(pd.Series)], axis=1)     



df_reviews2.head()                                                                                                            # Revisamos como quedo el dataframe

Se creó una columna con sólo NaN, por lo que la eliminaremos

del df_reviews2[0]

Validamos que se haya almacenado correctamente la informacion al nuevo dataframe

fila_etiqueta = df_review.loc[1564]                                 # Datos en el df original
print(fila_etiqueta)
contenido_reviews = df_review.loc[1564, 'reviews']
print(contenido_reviews)

fila_etiqueta = df_reviews2.loc[1564]                               # Datos en el df ya desanidado
fila_etiqueta

#### Transformación de datos

Vemos que tipo de dato es cada variable, la cantidad de registros y columnas

df_reviews2.info()

Calculamos la cantidad de valores nulos en el DataFrame df_reviews2

Calculamos la cantidad de registros totalmente nulos, y procedemos a eliminarlos

cantidad_filas_nulas = df_reviews2.isnull().all(axis=1).sum()
print("Cantidad de filas con todos los valores nulos:", cantidad_filas_nulas)
if cantidad_filas_nulas > 0:
    df_reviews2 = df_reviews2.dropna(how='all').reset_index(drop=True)
    print('Se eliminaron los registros totalmente vacios')
else:
    print('No se eliminaron registros, ya que no se encontraron registros totalmente vacios')

Luego buscamos cuantos nulos quedan en nuestro dataframe

nulos = t.PorcentajeNulos(df_reviews2)
nulos

Eliminamos los registros nulos, ya que no representan más del 4% del total 

df_reviews2 = df_reviews2.dropna()

La columna funny tiene un 86% de celdas vacías (no nulas), por lo que procedemos a eliminarla del df

vacios = pd.DataFrame(df_reviews2['funny'].value_counts() / len(df_reviews2) * 100)
vacios = vacios.rename(columns={'count': '%Vacios'})
vacios

df_reviews2 = df_reviews2.drop('funny', axis=1)                 # Eliminamos la columna

Sucede lo mismo con last_edited

vacios = pd.DataFrame(df_reviews2['last_edited'].value_counts() / len(df_reviews2) * 100)
vacios = vacios.rename(columns={'count': '%Vacios'})
vacios

df_reviews2 = df_reviews2.drop('last_edited', axis=1)           # Eliminamos la columna

Revisamos los duplicados

filas_duplicadas = df_reviews2[df_reviews2.duplicated(keep='first')]
filas_duplicadas.head()

Al validar que sean resultados duplicados, observamos que realmente no son registros duplicados por lo que no se eliminarán.

user_id = 'ImSeriouss'

fila_usuario = df_reviews2.loc[df_reviews2['user_id'] == user_id]
fila_usuario.head()

Para simplificar el dataframe, agregaremos la columna year_posted la cual contiene únicamente el año, para su posterior análisis

df_reviews2['year_posted'] = df_reviews2['posted'].str.extract(r'(\d{4})')          # Extraemos el año
mediana = df_reviews2['year_posted'].dropna().astype(float).median()                  # Como son más de 10,000 datos que no cuentan con el dato del año
df_reviews2['year_posted'] = df_reviews2['year_posted'].fillna(mediana)               # Rellenamos esos nulos con la media de los años con de los que si tienen el dato del year_posted

### Feature Engineering

#### Analisis de sentimientos

import nltk
from nltk.sentiment import SentimentIntensityAnalyzer
from tqdm.notebook import tqdm
import langid

sia = SentimentIntensityAnalyzer()

Revisamos el idioma en que están escritos los reviews 

contador_idiomas = {}

for i in df_reviews2['review']:                                         # En cada linea del dataframe
    idioma,_ = langid.classify(i)                                       
    contador_idiomas[idioma] = contador_idiomas.get(idioma,0) + 1       # Clasificar el idioma, agregar 1 al contador de dicho idioma y actualizar el diccionario

contador_idiomas

Para facilitar la visualización pasamos el diccionario a un dataframe, y calculamos el porcentaje que representa cada idioma sobre el total de reviews

contador_idiomas_sorted = dict(sorted(contador_idiomas.items(), key=lambda item: item[1], reverse=True))    # Acomodamos el diccionario de mayor a menor
pd_idiomas = pd.DataFrame(list(contador_idiomas_sorted.items()), columns=['Idioma', 'Conteo'])              # Guardamos el diccionario en un dataframe
pd_idiomas['Proporcion'] = round(pd_idiomas['Conteo'] *100 / pd_idiomas['Conteo'].sum().sum(),2)            # Calculamos el % que representa cada idioma
pd_idiomas

Podemos observar que el 85% de los reviews son en inglés, por lo que por fines de optimización de tiempos en este análisis consideraremos que todos los reviews están en inglés

##### VADER

res = {}                                                                    # Isntanciamos un diccionario vacio para almacenar los resultados
for i, linea in tqdm(df_reviews2.iterrows(), total=len(df_reviews2)):       # Iteramos en todas las lineas del df 
    text = linea['review']
    myid = linea['user_id']
    res[myid] = sia.polarity_scores(text)                                   # Determinamos su polarity score
 

vaders = pd.DataFrame(res).T                                                                # Resultados del análisis, que tan negativo, positivo o neutral es el review

merged_df = vaders.merge(df_reviews2, left_index=True, right_on='user_id', how='left')      # Unimos los resultados de VADER con el dataframe de reviews
merged_df

Validamos que se haya realizado correctamente el merge

fila = df_reviews2.loc[df_reviews2['user_id'] == 'js41637']
fila

Hacemos una nueva ponderación sobre las calificaciones resultantes del análisis de sentimientos, usamos la columna compound que es un "indicador global" (que considera el score pos, neg y neu) del sentimiento de ese review, y le asignamos nuevos valores:
* 0 si es Malo
* 1 si es Neutral
* 2 si es Bueno

merged_df['score'] = 1                                                              # Ponemos por default toda la columna con 1


for index, row in merged_df.iterrows():                                             # Iteramos sobre las filas 
    if row['compound'] > 0:                                                         # Actualizamos la columna 'score' basada en la puntuación compuesta ('compound')
        merged_df.at[index, 'score'] = 2
    elif row['compound'] < 0:
        merged_df.at[index, 'score'] = 0


merged_df

Validamos que haya funcionado correctamente

score = merged_df.loc[merged_df['score'] == 1]
score

Eliminamos las columnas que ya no se necesitan: neg, neu, pos, compound, posted, review

eliminar = ['neg','neu','pos','compound','review','posted', 'user_url', 'helpful']
merged_df = merged_df.drop(columns=eliminar)

Vemos como queda el dataframe

merged_df.head()

Modificamos la columna recommend para facilitar el análisis, donde  
False = 0  
True = 1

Validamos antes del cambio, cuantos True y False hay

merged_df['recommend'].value_counts()

Realizamos el cambio

merged_df['recommend'] = merged_df['recommend'].replace({False: 0, True: 1})
merged_df

Validamos que no se haya modificado la cantidad de True y False

merged_df['recommend'].value_counts()

Ya que contamos con el dataframe limpio, procedemos con la modificación de los tipos de datos de cada columna que haga falta

merged_df['user_id'] = merged_df['user_id'].astype(str)
merged_df['year_posted'] = merged_df['year_posted'].astype(int)

Guardamos el conjunto ya limpio a un csv

merged_df.to_csv('users_reviews_cleaned.csv', index = False, encoding='utf-8')

Este es el fin de este ETL, porfavor da click [aqui](02_EDA.ipynb) para continuar con el EDA.
