==> 03_dfAuxiliares.ipynb <==
### Preparacion de dataframes auxiliares para los endpoints

En este notebook, estaremos preparando dataframes auxiliares que contienen únicamente información correspondiente a la consulta en particular, y a su vez estaremos desarrollando y probando las funciones que serán los endpoints en la API.

#### Importamos las librerias a utilizar

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import Tools as t
import pyarrow.parquet as pq
warnings.filterwarnings("ignore")

Cargamos los datasets que ya previamente limpiados

games = pd.read_csv('data_ok/games_cleaned.csv', encoding='utf-8')
items = pd.read_csv('data_ok/users_items_cleaned.csv', encoding='utf-8')
reviews = pd.read_csv('data_ok/users_reviews_cleaned.csv', encoding='utf-8')

 <h2 align=center> ______________________________________________________________________<h2>

#### PlayTimeGenre. Preparamos el dataframe para la función 

##### Procesamiento

Unimos los dataframes que contienen los campos necesarios para realizar la consulta

df_PlayTimeGenre = pd.merge(items, games, on = 'item_id', how= 'left')  
eliminar = ['item_name','user_id','items_count','steam_id','app_name', 'developer']
df_PlayTimeGenre = df_PlayTimeGenre.drop(columns=eliminar)
df_PlayTimeGenre

Revisamos los nulos, de ser necesario, los eliminamos

nulos = t.PorcentajeNulos(df_PlayTimeGenre)
nulos

df_PlayTimeGenre.dropna()

Buscamos los registros con playtiem_forever == 0, quiere decir que no jugaron nunca y procedemos a eliminarlos

df_PlayTimeGenre = df_PlayTimeGenre.loc[df_PlayTimeGenre['playtime_forever'] != 0]      
df_PlayTimeGenre 

Agrupamos por géneros y por años, y calculamos el total de horas jugadas

df_PlayTimeGenre = df_PlayTimeGenre.groupby([ 'genres', 'released_year'])['playtime_forever'].sum().reset_index()               # Hacemos un dataframe con las columnas necesarias ya argupadas, y suma el total de horas jugadas
df_PlayTimeGenre.rename(columns= {'playtime_forever':'Total_horas_jugadas'}, inplace = True)                                    # Renombramos las columnas
df_PlayTimeGenre.head()


Guardamos el dataframe en un parquet, esto para minimizar el peso del archivo y facilitar la consulta del API

df_PlayTimeGenre.to_parquet('datos_parquet/df_PlayTimeGenre.parquet')

##### Función

def PlayTimeGenre(df, genero:str):

    mascara = (df['genres'] == genero)                                                                                         # Filtramos el dataframe para que solo sean de el género de la consulta
    df_PlayTimeGenre_sorted = df[mascara].sort_values(by='Total_horas_jugadas', ascending=False).head(1)                       # Acomodamos el df en orden descendente en relacion a Total horas jugadas

    year = df_PlayTimeGenre_sorted['released_year'].values[0]                                                                  # Sacamos el año del género con más horas jugadas

    resultado = {f'Año de lanzamiento con más horas jugadas para el género {genero}': year}
    return resultado

Probamos la función

prueba = PlayTimeGenre(df_PlayTimeGenre,'Action')
print(prueba)

 <h2 align=center> ______________________________________________________________________<h2>

#### UserForGenre. 

##### Procesamiento

Unimos los dataframes que contienen los campos necesarios para realizar la consulta

df_UserForGenre = pd.merge(items, games, on = 'item_id', how= 'left')  
eliminar = ['item_id', 'item_name','items_count','steam_id','app_name', 'developer']
df_UserForGenre = df_UserForGenre.drop(columns=eliminar)

df_UserForGenre

Buscamos los registros con playtiem_forever == 0, quiere decir que no jugaron nunca y procedemos a eliminarlos

df_UserForGenre = df_UserForGenre.loc[df_UserForGenre['playtime_forever'] != 0]      
df_UserForGenre                  

Revisamos la cantidad de nulos que se generaron

nulos =  t.PorcentajeNulos(df_UserForGenre)
nulos

Eliminamos los nulos

df_UserForGenre1 = df_UserForGenre.dropna()

Agrupamos por user_id, género y año el dataframe, para sumar el total de horas jugadas

df_UserForGenre_group = df_UserForGenre1.groupby(['genres', 'user_id', 'released_year'])['playtime_forever'].sum().reset_index()                # Hacemos un dataframe con las columnas necesarias ya argupadas, y suma el total de horas jugadas
df_UserForGenre_group.rename(columns= {'playtime_forever':'Total_horas_jugadas'}, inplace = True)                                               # Le cambiamos el nombre a la columna

Almacenamos el dataframe en un archivo parquet

df_UserForGenre_group.to_parquet('datos_parquet/df_UserForGenre.parquet')

##### Función

Ya con el dataframe listo, procedemos a realizar la función 

def UserForGenre(df, genero: str):                     
    mascara = (df['genres'] == genero)
    df_UserForGenre = df[mascara]

    df_UserForGenre_sorted = df_UserForGenre.groupby('user_id')['Total_horas_jugadas'].sum().reset_index()                        # Obtenemos el usuario con más horas jugadas para el género
    user_max_playtime = df_UserForGenre_sorted.sort_values(by='Total_horas_jugadas', ascending=False).iloc[0]['user_id']          # Acomodamos de mayor a menor el dataframe, y tomamos el primero

    df_UserGenreUser = df_UserForGenre[df_UserForGenre['user_id'] == user_max_playtime]                                           # Filtramos los datos para ese usuario y género

    df_UserGenreUserByYear = df_UserGenreUser.groupby('released_year')['Total_horas_jugadas'].sum().reset_index()                 # Obtenemos el total de horas jugadas por año para el usuario y género seleccionados

    horas_por_anio = df_UserGenreUserByYear.to_dict(orient='records')                                                             # Creamos la lista de años y horas jugadas

    return {f"Usuario con más horas jugadas para Género {genero}": user_max_playtime,"Horas jugadas": horas_por_anio}



Probamos la función

resultado = UserForGenre(df_UserForGenre_group, 'Action')
print(resultado)

 <h2 align=center> ______________________________________________________________________<h2>

#### UsersRecommend. 

##### Procesamiento

Unimos los dataframes que contienen los campos necesarios para realizar la consulta

mascara = (reviews['recommend'] == 1) & ((reviews['score'] == 1) | (reviews['score'] == 2))
best_reviews = reviews[mascara]

df_best_reviews_count = best_reviews.groupby(['item_id', 'year_posted']).size().reset_index(name='Total_good_reviews')         # Obtenemos el conteo de malas reseñas por 'item_id' y 'year_posted'

df_best_reviews_2 = pd.merge(df_best_reviews_count,games)                                                                      # Unimos el dataframe creado, con el de games para obtener el título de los videojuegos
eliminar = ['genres','released_year','developer']                                                                              # Para este endpoint no son necesarias estas columnas
df_UsersRecommend = df_best_reviews_2.drop(columns=eliminar)                                                                   # Por lo que las eliminamos
df_UsersRecommend

Guardamos el dataframe en un archivo parquet

df_UsersRecommend.to_parquet('datos_parquet/df_UsersRecommend.parquet')

##### Función

Hacemos la funcion UsersRecommend

def UsersRecommend(df_best_reviews_2,anio:int):

    mascara = (df_best_reviews_2['year_posted'] == anio)   
    df_best_reviews_3 = df_best_reviews_2[mascara]
    name_counts = df_best_reviews_3['app_name'].value_counts().head(3)
 
    resultados = []
    for puesto, (name, count) in enumerate(name_counts.items(), start=1):                            # Crear una lista de diccionarios con el formato solicitado
        resultados.append({f"Puesto {puesto}": name})

    return resultados

Probamos la función

resultado_bestrev = UsersRecommend(df_UsersRecommend,2014)
print(resultado_bestrev)

 <h2 align=center> ______________________________________________________________________<h2>

#### UsersWorstDeveloper. Preparamos el dataframe para la funcion 

##### Procesamiento

mascara = (reviews['score'] == 0) & (reviews['recommend'] == 0)                                                                 # Para ser considerado malo, debe tener recommend = o y score = 0
worst_reviews = reviews[mascara]                                                                                                # Filtramos el df con estas condiciones

df_worst_reviews_count = worst_reviews.groupby(['item_id', 'year_posted']).size().reset_index(name='Total_bad_reviews')         # Obtenemos el conteo de malas reseñas por 'item_id' y 'year_posted'

df_worst_reviews_2 = pd.merge(df_worst_reviews_count,games)
eliminar = ['genres','released_year','app_name']                                                                                # Para este endpoint no son necesarias estas columnas
df_worst_reviews_2 = df_worst_reviews_2.drop(columns=eliminar)                                                                  # Por lo que las eliminamos
df_worst_reviews_2

df_worst_reviews_2.to_parquet('datos_parquet/df_UsersWorstDeveloper.parquet')

##### Función

Hacemos la funcion UsersWorstDeveloper

def UsersWorstDeveloper(df_worst_reviews_2,anio:int):

    mascara = (df_worst_reviews_2['year_posted'] == anio)   
    df_worst_reviews_3 = df_worst_reviews_2[mascara]
    developer_counts = df_worst_reviews_3['developer'].value_counts().head(3)
 
    resultados = []
    for puesto, (developer, count) in enumerate(developer_counts.items(), start=1):                            # Crear una lista de diccionarios con el formato solicitado
        resultados.append({f"Puesto {puesto}": developer})

    return resultados

Lo probamos

resultado_final = UsersWorstDeveloper(df_worst_reviews_2,2015)
print(resultado_final)

 <h2 align=center> ______________________________________________________________________<h2>

#### Sentiment Analysis. Preparamos el dataframe para la función

##### Procesamiento

df_Sentiment_An = pd.merge(games[['item_id','developer']], reviews[['item_id','score']], on ='item_id', how = 'inner')                          # Unimos el df de games (solo la columna developer) con el df reviews (solo la columna score)
df_Sentiment_An.drop(columns=['item_id'], inplace=True)                                                                                         # No necesitamos la columna item_id para esta consulta
df_Sentiment_An.head()

Guardamos el dataframe en un archivo parquet para su posterior analisis

df_Sentiment_An.to_parquet('datos_parquet/df_SentimentAnalysis.parquet')

##### Función

def SentimentAnalysis(df, developer:str):
    
    mascara = (df['developer'] == developer)
    df1 = df[mascara]

    pos = df1[df1['score'] == 2]['score'].sum()
    neu = df1[df1['score'] == 1]['score'].sum()
    neg = df1[df1['score'] == 0]['score'].sum()

    final = {f'{developer}:[Negative = {neg}, Neutral = {neu}, Positive = {pos}]'}

    return final

prueba = SentimentAnalysis(df_Sentiment_An,'Stainless Games Ltd')
prueba

Este es el fin de esta parte del proceso, porfavor da click [aqui](04_SistemaDeRecomendacion.ipynb) para continuar.
