==> 04_SistemaDeRecomendacion.ipynb <==
## Sistema de Recomendación


### En este caso utilizaremos la Similaridad del Coseno

Importamos las librerias

import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
from sklearn.neighbors import NearestNeighbors
import matplotlib.pyplot as plt
import Tools as t
from sklearn.metrics.pairwise import cosine_similarity
import pyarrow.parquet as pq
import sklearn

Leemos los datos y los guardamos en un dataframe

df_recomendacion = pd.read_csv('data_ok/df_recomendacion.csv', encoding = 'utf-8')

Revisamos la información

df_recomendacion.head()

Creamos una matriz User-Videojuego, para aplicar la Similaridad del Coseno

matriz = df_recomendacion.pivot_table(index='app_name', columns='user_id', values='calificaciones')                     # Hacemos la matriz
matriz.head(20)

Normalizamos los valores de calificación restandoas con su promedio, para que quede en un rango de valores más equitativo 

matriz_normal = matriz.subtract(matriz.mean(axis=1), axis = 'rows')             
matriz_normal.head()

Rellenamos todos los valores nulos con 0, de lo contrario no podremos utilizar similarity coseno

matriz_normal = matriz_normal.fillna(0)                                                                      # Rellenamos con 0

matriz_normal = matriz_normal.loc[:, (matriz_normal != 0).any(axis=0)]                                       # Eliminamos los elementos completamente nulos

item_similarity_coseno = cosine_similarity(matriz_normal)                                                    # Aplicamos la similaridad del coseno
item_similarity_coseno

Para mayor facilidad y visualización, pasamos el array item_similarity_coseno a un dataframe de pandas

df_i_sim_cos= pd.DataFrame(item_similarity_coseno, index = matriz_normal.index, columns = matriz_normal.index)
df_i_sim_cos

Guardamos el dataframe en un archivo parquet para su posterior análisis

df_i_sim_cos.to_parquet('datos_parquet/df_recomendacion_juego.parquet')

### Funcion

def recomendacion_juego(juego):

    similares = df_i_sim_cos.sort_values(by=juego, ascending=False).index[1:6]                               # Acomoda los valores de mayor a menor, y toma sólo los primeros 5
    print(f'Los 5 juegos más similares a {juego} son:\n')
    for count, item in enumerate(similares, start=1):                                                        # Itera sobre el dataframe similares para imprimir los resultados
        print(f'No. {count}: {item}')

Vamos a probar la función

top = recomendacion_juego('100% Orange Juice')

top = recomendacion_juego('theHunter Classic')

 <h2 align=center> ______________________________________________________________________<h2>

Este es el fin del procesamiento de los datasets. Puede consultar el archivo [main.py](main.py) para observar las funciones que se deployarán, o bien el archivo [requirements.txt](requirements.txt) que fue utilizado para hacer el deploy del API en render.

Después de este punto puede acceder [aqui](https://steam-pi-4c9l.onrender.com/docs#/) para probar las funciones previamente mencionadas en estos notebooks.

O bien puede acceder al video resumen del proyecto [aqui](https://www.youtube.com/watch?v=3FRkXWo55Yw).

Gracias.
